src/edu/ycp/cs320/ms1/model/User.java
package edu.ycp.cs320.ms1.model;

public class User {
	private int userId;
	private String lastname;
	private String firstname;
	
	public User() {
		
	}
	
	public void setUserId(int userId) {
		this.userId = userId;
	}
	
	public int getUserId() {
		return userId;
	}
	
	public void setLastname(String lastname) {
		this.lastname = lastname;
	}
	
	public String getLastname() {
		return lastname;
	}
	
	public void setFirstname(String firstname) {
		this.firstname = firstname;
	}
	
	public String getFirstname() {
		return firstname;
	}
}
